+++ generator/streets/streets.cpp
@@ -16,7 +16,7 @@ void GenerateStreets(std::string const & pathInRegionsIndex, std::string const &
                      std::string const & pathInStreetsTmpMwm,
                      std::string const & pathInGeoObjectsTmpMwm,
                      std::string const & pathOutStreetsKv,
-                     bool /*verbose*/, size_t threadsCount)
+                     bool /*verbose*/, unsigned int threadsCount)
 {
   LOG(LINFO, ("Start generating streets..."));
   auto timer = base::Timer();
